refactor(server): peel search routes into routers/search.py (router-split, round-5 #51 / R5-25 PR26)
Fifteenth router peel. Moves the two search surfaces into a leaf routers/search.py:
GET /api/search/all (cross-project federated semantic search + the Phase 16.2
API-boundary path sanitisation) and POST /api/search/query (structured query —
typed conditions AND/OR-combined via query_builder + optional semantic vector leg),
with StructuredQuery, _structured_sort_key, and _split_csv (its only caller is
search_all). The bulk-fetch helpers shared with the media group are imported from
mediarecords.py (the ONE shared instance); path sanitisation from pathres.py.

Behaviour-preserving: paths/methods/scopes unchanged (SPA + CLI unaffected).
server.py 1068 → 911. Route-level federation tests unaffected (they patch the
federation module, which the router reads); chat-peel guard updated (_split_csv
moved here).
